refactor(cli): Split 5 oversized files into directory modules (#336)
Convert 5 files exceeding 400 lines into directory modules following
the existing config/ pattern, enforcing one concept per file.

- http_client.rs (1327L) → http_client/ (11 files)
- start.rs (1035L) → start/ (5 files)
- collection.rs (700L) → collection/ (3 files)
- p2p.rs (510L) → p2p/ (5 files)
- acp.rs (425L) → acp/ (3 files)

Purely structural — no behavior changes. All tests pass.
