--- crates/cli/src/commands/client/acp.rs
@@ -1,425 +1,0 @@
-//! ACP (Access Control Policy) command implementation
-
-use std::path::PathBuf;
-
-use clap::{Args, Subcommand};
-
-use super::http_client::HttpClient;
-use super::{get_data_from_args, ClientContext};
-use crate::error::Result;
-
-/// Interact with Access Control Policies
-#[derive(Args, Debug)]
-pub struct AcpArgs {
-    #[command(subcommand)]
-    pub command: AcpCommand,
-}
-
-/// ACP subcommands
-#[derive(Subcommand, Debug)]
-pub enum AcpCommand {
-    /// Manage document-level ACP
-    Document(AcpDocumentArgs),
-    /// Manage node-level ACP
-    Node(AcpNodeArgs),
-}
-
-/// ACP document subcommands
-#[derive(Args, Debug)]
-pub struct AcpDocumentArgs {
-    #[command(subcommand)]
-    pub command: AcpDocumentCommand,
-}
-
-/// ACP document subcommands
-#[derive(Subcommand, Debug)]
-pub enum AcpDocumentCommand {
-    /// Manage ACP policies
-    Policy(AcpPolicyArgs),
-    /// Manage ACP relationships
-    Relationship(AcpDocumentRelationshipArgs),
-}
-
-/// ACP policy subcommands
-#[derive(Args, Debug)]
-pub struct AcpPolicyArgs {
-    #[command(subcommand)]
-    pub command: AcpPolicyCommand,
-}
-
-/// ACP policy subcommands
-#[derive(Subcommand, Debug)]
-pub enum AcpPolicyCommand {
-    /// Add a new ACP policy
-    Add(AcpPolicyAddArgs),
-}
-
-/// Arguments for acp policy add command
-#[derive(Args, Debug)]
-pub struct AcpPolicyAddArgs {
-    /// The policy definition (YAML or JSON format)
-    #[arg(value_name = "POLICY")]
-    pub policy: Option<String>,
-
-    /// Read policy from file
-    #[arg(long, short = 'f', value_name = "FILE")]
-    pub file: Option<PathBuf>,
-}
-
-/// ACP document relationship subcommands
-#[derive(Args, Debug)]
-pub struct AcpDocumentRelationshipArgs {
-    #[command(subcommand)]
-    pub command: AcpDocumentRelationshipCommand,
-}
-
-/// ACP document relationship subcommands
-#[derive(Subcommand, Debug)]
-pub enum AcpDocumentRelationshipCommand {
-    /// Add a document ACP relationship
-    Add(AcpDocumentRelationshipAddArgs),
-    /// Delete a document ACP relationship
-    Delete(AcpDocumentRelationshipDeleteArgs),
-}
-
-/// Arguments for acp document relationship add command
-#[derive(Args, Debug)]
-pub struct AcpDocumentRelationshipAddArgs {
-    /// Collection name
-    #[arg(long, short = 'c')]
-    pub collection: Option<String>,
-
-    /// Document ID
-    #[arg(long = "docID")]
-    pub doc_id: Option<String>,
-
-    /// Relation name
-    #[arg(long, short = 'r')]
-    pub relation: Option<String>,
-
-    /// Actor (target identity)
-    #[arg(long, short = 'a')]
-    pub actor: Option<String>,
-}
-
-/// Arguments for acp document relationship delete command
-#[derive(Args, Debug)]
-pub struct AcpDocumentRelationshipDeleteArgs {
-    /// Collection name
-    #[arg(long, short = 'c')]
-    pub collection: Option<String>,
-
-    /// Document ID
-    #[arg(long = "docID")]
-    pub doc_id: Option<String>,
-
-    /// Relation name
-    #[arg(long, short = 'r')]
-    pub relation: Option<String>,
-
-    /// Actor (target identity)
-    #[arg(long, short = 'a')]
-    pub actor: Option<String>,
-}
-
-/// ACP node subcommands
-#[derive(Args, Debug)]
-pub struct AcpNodeArgs {
-    #[command(subcommand)]
-    pub command: AcpNodeCommand,
-}
-
-/// ACP node subcommands
-#[derive(Subcommand, Debug)]
-pub enum AcpNodeCommand {
-    /// Manage node ACP relationships
-    Relationship(AcpNodeRelationshipArgs),
-    /// Show ACP status
-    Status(AcpNodeStatusArgs),
-    /// Disable node ACP
-    Disable(AcpNodeDisableArgs),
-    /// Re-enable node ACP
-    ReEnable(AcpNodeReEnableArgs),
-}
-
-/// ACP node relationship subcommands
-#[derive(Args, Debug)]
-pub struct AcpNodeRelationshipArgs {
-    #[command(subcommand)]
-    pub command: AcpNodeRelationshipCommand,
-}
-
-/// ACP node relationship subcommands
-#[derive(Subcommand, Debug)]
-pub enum AcpNodeRelationshipCommand {
-    /// Add a node ACP relationship
-    Add(AcpNodeRelationshipAddArgs),
-    /// Delete a node ACP relationship
-    Delete(AcpNodeRelationshipDeleteArgs),
-}
-
-/// Arguments for acp node relationship add command
-#[derive(Args, Debug)]
-pub struct AcpNodeRelationshipAddArgs {
-    /// Relation name
-    #[arg(long, short = 'r')]
-    pub relation: Option<String>,
-
-    /// Actor (target identity)
-    #[arg(long, short = 'a')]
-    pub actor: Option<String>,
-}
-
-/// Arguments for acp node relationship delete command
-#[derive(Args, Debug)]
-pub struct AcpNodeRelationshipDeleteArgs {
-    /// Relation name
-    #[arg(long, short = 'r')]
-    pub relation: Option<String>,
-
-    /// Actor (target identity)
-    #[arg(long, short = 'a')]
-    pub actor: Option<String>,
-}
-
-/// Arguments for acp node status command
-#[derive(Args, Debug)]
-pub struct AcpNodeStatusArgs {}
-
-/// Arguments for acp node disable command
-#[derive(Args, Debug)]
-pub struct AcpNodeDisableArgs {}
-
-/// Arguments for acp node re-enable command
-#[derive(Args, Debug)]
-pub struct AcpNodeReEnableArgs {}
-
-impl AcpArgs {
-    /// Execute the ACP command
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        match &self.command {
-            AcpCommand::Document(args) => args.execute(ctx).await,
-            AcpCommand::Node(args) => args.execute(ctx).await,
-        }
-    }
-}
-
-impl AcpDocumentArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        match &self.command {
-            AcpDocumentCommand::Policy(args) => args.execute(ctx).await,
-            AcpDocumentCommand::Relationship(args) => args.execute(ctx).await,
-        }
-    }
-}
-
-impl AcpPolicyArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        match &self.command {
-            AcpPolicyCommand::Add(args) => args.execute(ctx).await,
-        }
-    }
-}
-
-impl AcpPolicyAddArgs {
-    /// Execute the acp policy add command
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let policy = get_data_from_args(&self.policy, &self.file)?;
-
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let response = client.acp_add_policy(&policy).await?;
-        println!("{}", serde_json::to_string_pretty(&response)?);
-        Ok(())
-    }
-}
-
-impl AcpDocumentRelationshipArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        match &self.command {
-            AcpDocumentRelationshipCommand::Add(args) => args.execute(ctx).await,
-            AcpDocumentRelationshipCommand::Delete(args) => args.execute(ctx).await,
-        }
-    }
-}
-
-impl AcpDocumentRelationshipAddArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let collection = self.collection.as_deref().ok_or_else(|| {
-            crate::error::Error::MissingInput("--collection is required".to_string())
-        })?;
-        let doc_id = self
-            .doc_id
-            .as_deref()
-            .ok_or_else(|| crate::error::Error::MissingInput("--docID is required".to_string()))?;
-        let relation = self.relation.as_deref().ok_or_else(|| {
-            crate::error::Error::MissingInput("--relation is required".to_string())
-        })?;
-        let actor = self
-            .actor
-            .as_deref()
-            .ok_or_else(|| crate::error::Error::MissingInput("--actor is required".to_string()))?;
-
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let result = client
-            .acp_doc_relationship_add(collection, doc_id, relation, actor)
-            .await?;
-        println!("{}", serde_json::to_string_pretty(&result)?);
-        Ok(())
-    }
-}
-
-impl AcpDocumentRelationshipDeleteArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let collection = self.collection.as_deref().ok_or_else(|| {
-            crate::error::Error::MissingInput("--collection is required".to_string())
-        })?;
-        let doc_id = self
-            .doc_id
-            .as_deref()
-            .ok_or_else(|| crate::error::Error::MissingInput("--docID is required".to_string()))?;
-        let relation = self.relation.as_deref().ok_or_else(|| {
-            crate::error::Error::MissingInput("--relation is required".to_string())
-        })?;
-        let actor = self
-            .actor
-            .as_deref()
-            .ok_or_else(|| crate::error::Error::MissingInput("--actor is required".to_string()))?;
-
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let result = client
-            .acp_doc_relationship_delete(collection, doc_id, relation, actor)
-            .await?;
-        println!("{}", serde_json::to_string_pretty(&result)?);
-        Ok(())
-    }
-}
-
-impl AcpNodeArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        match &self.command {
-            AcpNodeCommand::Relationship(args) => args.execute(ctx).await,
-            AcpNodeCommand::Status(args) => args.execute(ctx).await,
-            AcpNodeCommand::Disable(args) => args.execute(ctx).await,
-            AcpNodeCommand::ReEnable(args) => args.execute(ctx).await,
-        }
-    }
-}
-
-impl AcpNodeRelationshipArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        match &self.command {
-            AcpNodeRelationshipCommand::Add(args) => args.execute(ctx).await,
-            AcpNodeRelationshipCommand::Delete(args) => args.execute(ctx).await,
-        }
-    }
-}
-
-impl AcpNodeRelationshipAddArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let relation = self.relation.as_deref().ok_or_else(|| {
-            crate::error::Error::MissingInput("--relation is required".to_string())
-        })?;
-        let actor = self
-            .actor
-            .as_deref()
-            .ok_or_else(|| crate::error::Error::MissingInput("--actor is required".to_string()))?;
-
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let result = client.nac_add_relationship(relation, actor).await?;
-        println!("{}", serde_json::to_string_pretty(&result)?);
-        Ok(())
-    }
-}
-
-impl AcpNodeRelationshipDeleteArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let relation = self.relation.as_deref().ok_or_else(|| {
-            crate::error::Error::MissingInput("--relation is required".to_string())
-        })?;
-        let actor = self
-            .actor
-            .as_deref()
-            .ok_or_else(|| crate::error::Error::MissingInput("--actor is required".to_string()))?;
-
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let result = client.nac_remove_relationship(relation, actor).await?;
-        println!("{}", serde_json::to_string_pretty(&result)?);
-        Ok(())
-    }
-}
-
-impl AcpNodeStatusArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let result = client.nac_status().await?;
-        println!("{}", serde_json::to_string_pretty(&result)?);
-        Ok(())
-    }
-}
-
-impl AcpNodeDisableArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let result = client.nac_disable().await?;
-        println!("{}", serde_json::to_string_pretty(&result)?);
-        Ok(())
-    }
-}
-
-impl AcpNodeReEnableArgs {
-    pub async fn execute(&self, ctx: &ClientContext) -> Result<()> {
-        let client = HttpClient::new(&ctx.url)?
-            .with_auth_token(ctx.auth_token.clone())
-            .with_verbose(ctx.verbose);
-
-        let result = client.nac_re_enable().await?;
-        println!("{}", serde_json::to_string_pretty(&result)?);
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_acp_policy_add_args_with_data() {
-        let args = AcpPolicyAddArgs {
-            policy: Some("test policy".to_string()),
-            file: None,
-        };
-        assert!(args.policy.is_some());
-        assert!(args.file.is_none());
-    }
-
-    #[test]
-    fn test_acp_policy_add_args_with_file() {
-        let args = AcpPolicyAddArgs {
-            policy: None,
-            file: Some(PathBuf::from("policy.yaml")),
-        };
-        assert!(args.policy.is_none());
-        assert!(args.file.is_some());
-    }
-}
