--- crates/cli/src/commands/start.rs
@@ -1,1035 +1,0 @@
-//! Start command implementation
-
-use std::net::SocketAddr;
-use std::sync::Arc;
-
-use clap::Args;
-use tokio::sync::mpsc;
-use tokio::task::JoinHandle;
-use tracing::{error, info, warn};
-
-use crate::config::{AcpDocumentType, Config, DatastoreType};
-use crate::error::{Error, Result};
-use identity::Identity;
-use storage::backends::{DurabilityMode, RedbStoreOptions};
-
-const DEV_MODE_BANNER: &str = r#"
-******************************************
-**     DEVELOPMENT MODE IS ENABLED      **
-** ------------------------------------ **
-**   if this is a production database   **
-** disable development mode and restart **
-**   or you may risk losing all data    **
-******************************************
-"#;
-
-/// Arguments for the start command
-#[derive(Args, Debug)]
-pub struct StartArgs {
-    /// List of peers to connect to
-    #[arg(long, value_delimiter = ',')]
-    pub peers: Option<Vec<String>>,
-
-    /// Specify the maximum number of retries per transaction
-    #[arg(long)]
-    pub max_txn_retries: Option<u32>,
-
-    /// Specify the datastore to use (supported: badger, redb, memory)
-    #[arg(long)]
-    pub store: Option<String>,
-
-    /// Specify the datastore value log file size (in bytes)
-    #[arg(long)]
-    pub valuelogfilesize: Option<u64>,
-
-    /// Listen addresses for the p2p network (formatted as a libp2p MultiAddr)
-    #[arg(long, value_delimiter = ',')]
-    pub p2paddr: Option<Vec<String>>,
-
-    /// Disable the peer-to-peer network synchronization system
-    #[arg(long)]
-    pub no_p2p: Option<bool>,
-
-    /// List of origins to allow for CORS requests
-    #[arg(long, value_delimiter = ',')]
-    pub allowed_origins: Option<Vec<String>>,
-
-    /// Path to the public key for TLS
-    #[arg(long)]
-    pub pubkeypath: Option<String>,
-
-    /// Path to the private key for TLS
-    #[arg(long)]
-    pub privkeypath: Option<String>,
-
-    /// Enables development mode features
-    #[arg(long)]
-    pub development: Option<bool>,
-
-    /// Skip generating an encryption key. Encryption at rest will be disabled.
-    #[arg(long)]
-    pub no_encryption: Option<bool>,
-
-    /// Disable telemetry reporting
-    #[arg(long)]
-    pub no_telemetry: Option<bool>,
-
-    /// Disable signing of commits
-    #[arg(long)]
-    pub no_signing: Option<bool>,
-
-    /// Default key type to generate new node identity
-    #[arg(long)]
-    pub default_key_type: Option<String>,
-
-    /// Skip generating a searchable encryption key
-    #[arg(long)]
-    pub no_searchable_encryption: Option<bool>,
-
-    /// Hex formatted private key used to authenticate with ACP.
-    ///
-    /// The key type is auto-detected from the key length:
-    /// - 64 bytes (128 hex chars) → Ed25519
-    /// - 32 bytes (64 hex chars) → secp256k1
-    #[arg(short = 'i', long)]
-    pub identity: Option<String>,
-
-    /// Retry intervals for the replicator (comma-separated seconds)
-    #[arg(long, value_delimiter = ',')]
-    pub replicator_retry_intervals: Option<Vec<u32>>,
-
-    /// Storage durability mode: "eventual" (default, matches Go) or
-    /// "immediate" (fsync every commit, safer against OS crashes)
-    #[arg(long)]
-    pub durability: Option<String>,
-}
-
-impl StartArgs {
-    /// Execute the start command
-    pub async fn execute(self, mut config: Config) -> Result<()> {
-        // Apply start-specific flags to config
-        self.apply_to_config(&mut config)?;
-
-        // Create config if it doesn't exist
-        config.create_if_missing()?;
-
-        // Show development mode banner
-        if config.development {
-            eprintln!("{}", DEV_MODE_BANNER);
-        }
-
-        // Parse user identity from --identity flag if provided
-        let user_identity = self.parse_user_identity()?;
-
-        // Start the node
-        let node = Node::new(config, user_identity).await?;
-        node.run().await
-    }
-
-    /// Parse the user identity from the --identity flag.
-    ///
-    /// The identity flag should contain a hex-encoded private key.
-    /// Key type is auto-detected from byte length:
-    /// - 64 bytes → Ed25519
-    /// - 32 bytes → secp256k1
-    fn parse_user_identity(&self) -> Result<Option<std::sync::Arc<identity::RawIdentity>>> {
-        let hex_key = match &self.identity {
-            Some(key) => key,
-            None => return Ok(None),
-        };
-
-        // Remove 0x prefix if present
-        let hex_str = hex_key.strip_prefix("0x").unwrap_or(hex_key);
-
-        // Decode hex to bytes
-        let key_bytes = hex::decode(hex_str).map_err(|e| {
-            Error::InvalidIdentity(format!("invalid hex in --identity flag: {}", e))
-        })?;
-
-        // Auto-detect key type from byte length
-        let key_type = match key_bytes.len() {
-            64 => identity::IdentityKeyType::Ed25519,
-            32 => identity::IdentityKeyType::Secp256k1,
-            n => {
-                return Err(Error::InvalidIdentity(format!(
-                    "invalid key length {} bytes: expected 64 (ed25519) or 32 (secp256k1)",
-                    n
-                )));
-            }
-        };
-
-        // Create identity from bytes
-        let raw_identity = identity::RawIdentity::from_identity_key_type(key_type, &key_bytes)?;
-
-        let did = raw_identity.did()?;
-        info!("User identity DID: {}", did);
-
-        Ok(Some(std::sync::Arc::new(raw_identity)))
-    }
-
-    /// Apply start command flags to config
-    ///
-    /// Returns an error if any flag value fails to parse.
-    pub fn apply_to_config(&self, config: &mut Config) -> Result<()> {
-        if let Some(ref peers) = self.peers {
-            config.net.peers = peers.clone();
-        }
-        if let Some(retries) = self.max_txn_retries {
-            config.datastore.max_txn_retries = retries;
-        }
-        if let Some(ref store) = self.store {
-            config.datastore.store = store.parse()?;
-        }
-        if let Some(size) = self.valuelogfilesize {
-            config.datastore.valuelogfilesize = size;
-        }
-        if let Some(ref addrs) = self.p2paddr {
-            config.net.p2p_addresses = addrs.clone();
-        }
-        if let Some(no_p2p) = self.no_p2p {
-            config.net.p2p_disabled = no_p2p;
-        }
-        if let Some(ref origins) = self.allowed_origins {
-            config.api.allowed_origins = origins.clone();
-        }
-        if let Some(ref path) = self.pubkeypath {
-            config.api.pubkey_path = path.clone();
-        }
-        if let Some(ref path) = self.privkeypath {
-            config.api.privkey_path = path.clone();
-        }
-        if let Some(dev) = self.development {
-            config.development = dev;
-        }
-        if let Some(no_enc) = self.no_encryption {
-            config.datastore.no_encryption = no_enc;
-        }
-        if let Some(no_tel) = self.no_telemetry {
-            config.telemetry_disabled = no_tel;
-        }
-        if let Some(no_sign) = self.no_signing {
-            config.datastore.no_signing = no_sign;
-        }
-        if let Some(ref key_type) = self.default_key_type {
-            config.datastore.default_key_type = key_type.clone();
-        }
-        if let Some(no_se) = self.no_searchable_encryption {
-            config.datastore.no_searchable_encryption = no_se;
-        }
-        if let Some(ref intervals) = self.replicator_retry_intervals {
-            config.replicator_retry_intervals = intervals.clone();
-        }
-        if let Some(ref durability) = self.durability {
-            config.datastore.durability = match durability.as_str() {
-                "immediate" => DurabilityMode::Immediate,
-                "eventual" => DurabilityMode::Eventual,
-                other => {
-                    return Err(Error::InvalidConfig(format!(
-                        "invalid durability mode '{}': expected 'immediate' or 'eventual'",
-                        other
-                    )));
-                }
-            };
-        }
-        Ok(())
-    }
-}
-
-/// Tracks spawned P2P background tasks for graceful shutdown.
-struct P2PTasks {
-    /// P2P host event loop task
-    host_task: JoinHandle<()>,
-    /// Replication loop task (processes incoming blocks)
-    replication_task: JoinHandle<()>,
-    /// Host event handler task (processes P2P events through coordinator)
-    event_handler_task: Option<JoinHandle<()>>,
-}
-
-/// DefraDB Node
-/// Node manages the DefraDB server lifecycle.
-#[doc(hidden)]
-pub struct Node {
-    config: Config,
-    p2p_handle: Option<p2p::P2PHostHandle>,
-    p2p_tasks: Option<P2PTasks>,
-    http_server: Option<defra_http::Server>,
-    /// Shutdown signal sender (for tests)
-    #[doc(hidden)]
-    pub shutdown_tx: mpsc::Sender<()>,
-    shutdown_rx: mpsc::Receiver<()>,
-    /// User identity from --identity flag (for ACP authentication).
-    /// Stored for future use in request context injection.
-    #[allow(dead_code)]
-    user_identity: Option<std::sync::Arc<identity::RawIdentity>>,
-}
-
-impl Node {
-    /// Create a new node
-    #[doc(hidden)]
-    pub async fn new(
-        config: Config,
-        user_identity: Option<std::sync::Arc<identity::RawIdentity>>,
-    ) -> Result<Self> {
-        info!("Initializing DefraDB node");
-        info!("Root directory: {}", config.rootdir.display());
-        info!("Data directory: {}", config.data_path().display());
-
-        // Initialize peer keypair from keyring (if P2P enabled and keyring not disabled)
-        let peer_keypair = if !config.net.p2p_disabled && !config.keyring.disabled {
-            Some(Self::init_peer_key(&config)?)
-        } else if !config.net.p2p_disabled {
-            info!("Keyring disabled, using ephemeral peer identity");
-            None
-        } else {
-            None
-        };
-
-        // Initialize storage, database, and set up P2P and HTTP server
-        let (p2p_handle, p2p_tasks, http_server) = match config.datastore.store {
-            DatastoreType::Memory => {
-                info!("Using in-memory datastore");
-                let store = Arc::new(storage::MemoryStore::new());
-                // Use in-memory ACP store for memory datastore
-                let acp_store: Arc<dyn acp::AcpStore> = Arc::new(acp::MemoryAcpStore::new());
-                let zanzibar_store: Arc<dyn acp::ZanzibarStore> =
-                    Arc::new(acp::MemoryZanzibarStore::new());
-                info!("Using in-memory ACP store");
-                Self::init_store_and_server(
-                    store,
-                    &config,
-                    peer_keypair,
-                    user_identity.clone(),
-                    acp_store,
-                    zanzibar_store,
-                )
-                .await?
-            }
-            DatastoreType::Badger => {
-                info!("Using Redb datastore at {}", config.data_path().display());
-                let opts = RedbStoreOptions::new().with_durability(config.datastore.durability);
-                let store = Arc::new(storage::RedbStore::open_with_options(
-                    config.data_path(),
-                    opts,
-                )?);
-                // Use unified ACP store backed by main database with namespace isolation
-                info!("Using unified ACP store (namespace isolated in main database)");
-                let acp_store: Arc<dyn acp::AcpStore> =
-                    Arc::new(acp::PersistentAcpStore::from_store(store.clone()));
-                let zanzibar_store: Arc<dyn acp::ZanzibarStore> =
-                    Arc::new(acp::PersistentZanzibarStore::from_store(store.clone()));
-                Self::init_store_and_server(
-                    store,
-                    &config,
-                    peer_keypair,
-                    user_identity.clone(),
-                    acp_store,
-                    zanzibar_store,
-                )
-                .await?
-            }
-        };
-
-        let (shutdown_tx, shutdown_rx) = mpsc::channel(1);
-
-        Ok(Self {
-            config,
-            p2p_handle,
-            p2p_tasks,
-            http_server: Some(http_server),
-            shutdown_tx,
-            shutdown_rx,
-            user_identity,
-        })
-    }
-
-    /// Initialize or load the peer key from keyring.
-    ///
-    /// If a peer key exists in the keyring, it is loaded and converted to a libp2p Keypair.
-    /// If no peer key exists, a new Ed25519 key is generated and stored in the keyring.
-    fn init_peer_key(config: &Config) -> Result<p2p::Keypair> {
-        use crate::config::KeyringBackend;
-        use keyring::{FileKeyring, Keyring, SystemKeyring, PEER_KEY};
-
-        let kr: Box<dyn Keyring> = match config.keyring.backend {
-            KeyringBackend::File => {
-                let path = if config.keyring.path.starts_with('/') {
-                    std::path::PathBuf::from(&config.keyring.path)
-                } else {
-                    config.rootdir.join(&config.keyring.path)
-                };
-                let secret =
-                    keyring::load_secret_from_env().map_err(|e| Error::Keyring(e.to_string()))?;
-                Box::new(
-                    FileKeyring::open(&path, secret).map_err(|e| Error::Keyring(e.to_string()))?,
-                )
-            }
-            KeyringBackend::System => Box::new(SystemKeyring::open(&config.keyring.namespace)),
-        };
-
-        // Try to load existing peer key
-        match kr.get(PEER_KEY) {
-            Ok(key_bytes) => {
-                info!("Loaded existing peer key from keyring");
-                Self::derive_and_log_identity_did(&key_bytes)?;
-                Self::keypair_from_ed25519_bytes(&key_bytes)
-            }
-            Err(keyring::Error::NotFound(_)) => {
-                info!("Generating new peer key");
-                use crypto::Key;
-                let private_key = crypto::generate_ed25519()
-                    .map_err(|e| Error::Keyring(format!("failed to generate peer key: {}", e)))?;
-                let key_bytes = private_key.raw();
-
-                // Store in keyring
-                kr.set(PEER_KEY, &key_bytes)
-                    .map_err(|e| Error::Keyring(e.to_string()))?;
-
-                Self::derive_and_log_identity_did(&key_bytes)?;
-                Self::keypair_from_ed25519_bytes(&key_bytes)
-            }
-            Err(e) => Err(Error::Keyring(e.to_string())),
-        }
-    }
-
-    /// Derive and log the node's DID from peer key bytes.
-    ///
-    /// Creates a RawIdentity from the key bytes, derives its DID, and logs it.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if identity creation or DID derivation fails, which
-    /// indicates corrupted key material or a crypto library failure.
-    fn derive_and_log_identity_did(key_bytes: &[u8]) -> Result<()> {
-        use identity::{Identity, IdentityKeyType, RawIdentity};
-
-        let identity = RawIdentity::from_identity_key_type(IdentityKeyType::Ed25519, key_bytes)?;
-        let did = identity
-            .did()
-            .map_err(|e| Error::Keyring(format!("failed to derive DID: {}", e)))?;
-        info!("Node identity DID: {}", did);
-        Ok(())
-    }
-
-    /// Convert Ed25519 key bytes to libp2p Keypair.
-    ///
-    /// Ed25519 keys are stored as 64 bytes: 32-byte seed + 32-byte public key.
-    /// libp2p expects the 32-byte seed to derive the keypair.
-    fn keypair_from_ed25519_bytes(key_bytes: &[u8]) -> Result<p2p::Keypair> {
-        use libp2p::identity::ed25519;
-
-        if key_bytes.len() != 64 {
-            return Err(Error::Keyring(format!(
-                "invalid peer key length: expected 64 bytes, got {}",
-                key_bytes.len()
-            )));
-        }
-
-        // Ed25519 key format: 32-byte seed + 32-byte public key
-        // libp2p needs the seed (first 32 bytes) to derive the keypair
-        let seed: [u8; 32] = key_bytes[..32]
-            .try_into()
-            .map_err(|_| Error::Keyring("invalid key format".to_string()))?;
-
-        let secret_key = ed25519::SecretKey::try_from_bytes(seed)
-            .map_err(|e| Error::Keyring(format!("invalid Ed25519 key: {}", e)))?;
-
-        Ok(p2p::Keypair::from(ed25519::Keypair::from(secret_key)))
-    }
-
-    /// Initialize store, database, P2P, and HTTP server.
-    ///
-    /// This function creates the database, loads collections, sets up the query
-    /// runner with proper transaction support, and returns the HTTP server.
-    ///
-    /// Returns a tuple of (P2PHostHandle, P2PTasks, HTTP Server) where the tasks
-    /// are tracked for graceful shutdown.
-    async fn init_store_and_server<S>(
-        store: Arc<S>,
-        config: &Config,
-        peer_keypair: Option<p2p::Keypair>,
-        user_identity: Option<std::sync::Arc<identity::RawIdentity>>,
-        acp_store: Arc<dyn acp::AcpStore>,
-        zanzibar_store: Arc<dyn acp::ZanzibarStore>,
-    ) -> Result<(
-        Option<p2p::P2PHostHandle>,
-        Option<P2PTasks>,
-        defra_http::Server,
-    )>
-    where
-        S: storage::corekv::Store + 'static,
-    {
-        // Extract DID from user identity for query runner (before consuming it)
-        // SECURITY: If user explicitly provides --identity, DID derivation MUST succeed.
-        // Failing silently to anonymous would violate user's security expectations.
-        let user_did = match &user_identity {
-            Some(identity) => Some(identity.did().map_err(|e| {
-                Error::InvalidIdentity(format!(
-                    "failed to derive DID from --identity flag: {}. \
-                     Verify your key is valid and matches --identity-key-type. \
-                     Remove --identity flag to run without authentication.",
-                    e
-                ))
-            })?),
-            None => None,
-        };
-
-        // Build database options with optional user identity
-        let mut db_options = db::DbOptions::new();
-        if let Some(identity) = user_identity {
-            db_options = db_options.with_node_identity_arc(identity);
-            info!("Database configured with user identity");
-        }
-
-        // Open database and load collections from storage
-        let mut database = db::DB::open_from_arc_with_options(store.clone(), db_options)
-            .await
-            .map_err(|e| Error::Storage(storage::Error::Other(e.to_string())))?;
-
-        // Create and configure event bus for GraphQL subscriptions
-        let event_bus: Arc<dyn events::Bus> = Arc::new(events::ChannelBus::new());
-        database.set_event_bus(event_bus.clone());
-        info!("Event bus configured for subscriptions");
-
-        // Now wrap database in Arc
-        let database = Arc::new(database);
-
-        let collection_count = database
-            .list_collections()
-            .map_err(|e| Error::Storage(storage::Error::Other(e.to_string())))?
-            .len();
-        info!("Loaded {} collection schema(s)", collection_count);
-
-        // Set up P2P if enabled
-        // Clone store before potential move for sync coordinator blockstore
-        let store_for_sync = store.clone();
-        let (p2p, mut p2p_events, p2p_host_task) = if config.net.p2p_disabled {
-            (None, None, None)
-        } else {
-            info!("Initializing P2P network");
-            let blockstore = Arc::new(blockstore::DefraBlockstore::new(store, false));
-            let bitswap_store = p2p::BitswapStoreAdapter::new(blockstore);
-            let (handle, events, host_task) = Self::start_p2p(
-                config,
-                bitswap_store,
-                peer_keypair,
-                config.net.pubsub_enabled,
-            )
-            .await?;
-            (Some(handle), Some(events), Some(host_task))
-        };
-
-        // Create HTTP server with database-backed query runner
-        let (http_server, p2p_tasks) = {
-            let api_address: SocketAddr =
-                config
-                    .api
-                    .address
-                    .parse()
-                    .map_err(|e: std::net::AddrParseError| {
-                        Error::InvalidApiAddress(config.api.address.clone(), e.to_string())
-                    })?;
-
-            let server_config = defra_http::ServerConfig {
-                address: api_address,
-                allowed_origins: config.api.allowed_origins.clone(),
-            };
-
-            // Create auto-committing fetcher for non-transactional queries
-            let fetcher = db::AutoCommitFetcher::new(database.clone());
-
-            // Create sync coordinator if P2P is enabled (shared between mutator and P2P adapter)
-            // Also captures task handles for graceful shutdown
-            let (sync_coordinator, replication_task, event_handler_task) =
-                if let Some(ref p2p_handle) = p2p {
-                    let sync_blockstore = Arc::new(blockstore::DefraBlockstore::new(
-                        store_for_sync.clone(),
-                        false,
-                    ));
-
-                    // Clone blockstore for merge handler (before moving into coordinator)
-                    let merge_blockstore = sync_blockstore.clone();
-
-                    // Create persistent collection store for P2P subscriptions
-                    let collection_store: Arc<dyn p2p::sync::P2PCollectionStorage> =
-                        Arc::new(p2p::sync::P2PCollectionStore::new(store_for_sync));
-
-                    let (coordinator, sync_events) =
-                        p2p::sync::SyncCoordinator::with_collection_store(
-                            p2p_handle.clone(),
-                            sync_blockstore,
-                            p2p::sync::SyncConfig::default(),
-                            collection_store,
-                        )
-                        .await
-                        .map_err(Error::P2P)?;
-
-                    let coordinator = Arc::new(coordinator);
-
-                    // Load persisted P2P collection subscriptions
-                    match coordinator.load_p2p_collections().await {
-                        Ok(count) => {
-                            if count > 0 {
-                                info!("Loaded {} persisted P2P collection subscription(s)", count);
-                            }
-                        }
-                        Err(e) => {
-                            warn!("Failed to load persisted P2P collections: {}", e);
-                        }
-                    }
-
-                    // Create merge handler for CRDT merging
-                    let merge_handler =
-                        Arc::new(db::DbMergeHandler::new(database.clone(), merge_blockstore));
-
-                    // Spawn replication loop to process incoming blocks
-                    // Track the task handle for graceful shutdown
-                    let coordinator_for_replication = coordinator.clone();
-                    let replication_config = p2p::sync::ReplicationConfig {
-                        continue_on_error: true,
-                        rebroadcast_on_merge: false, // Don't re-broadcast during initial sync
-                    };
-                    let replication_task = tokio::spawn(async move {
-                        info!("Starting replication loop for P2P sync");
-                        p2p::sync::ReplicationLoop::run(
-                            coordinator_for_replication,
-                            sync_events,
-                            merge_handler,
-                            replication_config,
-                        )
-                        .await;
-                        info!("Replication loop stopped");
-                    });
-
-                    // Spawn host event handler to process incoming P2P events through coordinator
-                    // Track the task handle for graceful shutdown
-                    let event_handler_task = if let Some(mut events) = p2p_events.take() {
-                        let coordinator_for_events = coordinator.clone();
-                        Some(tokio::spawn(async move {
-                            while let Some(event) = events.recv().await {
-                                // Log events for visibility
-                                match &event {
-                                    p2p::HostEvent::PeerConnected(peer) => {
-                                        info!("Peer connected: {}", peer);
-                                    }
-                                    p2p::HostEvent::PeerDisconnected(peer) => {
-                                        info!("Peer disconnected: {}", peer);
-                                    }
-                                    p2p::HostEvent::Listening(addr) => {
-                                        info!("Now listening on: {}", addr);
-                                    }
-                                    p2p::HostEvent::GossipMessage {
-                                        propagation_source,
-                                        topic,
-                                        ..
-                                    } => {
-                                        info!(
-                                            "Received gossip message on {} from {}",
-                                            topic, propagation_source
-                                        );
-                                    }
-                                    p2p::HostEvent::TwoStreamRequest { peer_id, request } => {
-                                        info!(
-                                            peer_id = %peer_id,
-                                            message_id = %request.metadata.message_id,
-                                            doc_id = %request.doc_id,
-                                            "Processing TwoStreamRequest through coordinator"
-                                        );
-                                    }
-                                    _ => {}
-                                }
-
-                                // Process event through coordinator for response handling
-                                if let Err(e) =
-                                    coordinator_for_events.handle_host_event(event).await
-                                {
-                                    error!("Failed to handle host event: {}", e);
-                                }
-                            }
-                        }))
-                    } else {
-                        None
-                    };
-
-                    info!("P2P sync coordinator initialized");
-                    (
-                        Some(coordinator),
-                        Some(replication_task),
-                        event_handler_task,
-                    )
-                } else {
-                    (None, None, None)
-                };
-
-            // Create mutator - use BroadcastMutator if P2P is enabled for network propagation
-            let mutator: Arc<dyn query::mutator::DocMutator> =
-                if let Some(ref coordinator) = sync_coordinator {
-                    Arc::new(db::BroadcastMutator::new(
-                        database.clone(),
-                        coordinator.clone(),
-                    ))
-                } else {
-                    Arc::new(db::AutoCommitMutator::new(database.clone()))
-                };
-
-            // Create transaction registry for explicit transaction support
-            let registry = db::DbTransactionRegistry::new(database.clone());
-
-            // Create collection provider for on-demand schema resolution
-            // This ensures newly added schemas are immediately available for queries
-            let collection_provider: Arc<dyn query::CollectionProvider> =
-                db::DbCollectionProvider::new_arc(database.clone());
-            info!(
-                "Collection provider configured ({} collection(s) available)",
-                database.list_collections().map(|c| c.len()).unwrap_or(0)
-            );
-
-            // Create LocalDocumentACP with the provided store
-            let acp_store_for_http = acp_store.clone();
-            let document_acp: Arc<dyn acp::DocumentACP> =
-                Arc::new(acp::LocalDocumentACP::new(acp_store));
-            info!("Document ACP configured");
-
-            // Create query runner with transaction, mutation, and ACP support
-            // Use the collection provider for on-demand schema resolution
-            let mut query_runner = query::QueryRunner::with_registry_and_provider(
-                fetcher,
-                collection_provider,
-                registry,
-            )
-            .with_mutator(mutator)
-            .with_acp(document_acp)
-            .with_lens_store(database.lens_store().clone());
-
-            // Wire default identity for ACP permission checks (from --identity CLI flag)
-            if let Some(did) = user_did {
-                info!("Query runner configured with default identity for ACP");
-                query_runner = query_runner.with_default_identity(did);
-            }
-
-            let runner = Arc::new(query_runner);
-
-            // Create REST operations that wrap the query runner
-            let rest_ops = query::rest::RestOperationsImpl::new(Arc::clone(&runner));
-
-            // Create HTTP server with REST endpoints enabled
-            // Cast the Arc<QueryRunner> to Arc<dyn QueryExecutor> for the server
-            let executor: Arc<dyn query::executor::QueryExecutor> = runner;
-            let mut server = defra_http::Server::from_arc_with_config(executor, server_config)
-                .with_rest(rest_ops);
-
-            // Wire P2P to HTTP server if enabled
-            if let Some(ref p2p_handle) = p2p {
-                let p2p_adapter = if let Some(ref coordinator) = sync_coordinator {
-                    // Use sync coordinator for replicator operations (enables auto-subscribe)
-                    // Also provide collection lookup so we can resolve names to CollectionIDs
-                    // for topic subscription (matching Go DefraDB behavior)
-                    let collection_lookup =
-                        crate::p2p_adapter::DbCollectionLookup::new_arc(database.clone());
-                    crate::p2p_adapter::P2PAdapter::with_sync_coordinator_and_lookup_arc(
-                        p2p_handle.clone(),
-                        coordinator.clone(),
-                        collection_lookup,
-                    )
-                } else {
-                    crate::p2p_adapter::P2PAdapter::new_arc(p2p_handle.clone())
-                };
-                server = server.with_p2p_arc(p2p_adapter);
-                info!("P2P HTTP endpoints enabled");
-            }
-
-            // Wire schema operations to HTTP server
-            let schema_adapter = crate::schema_adapter::SchemaAdapter::new_arc(database.clone());
-            server = server.with_schema_arc(schema_adapter);
-            info!("Schema HTTP endpoint enabled");
-
-            // Wire lens operations to HTTP server
-            match crate::lens_adapter::LensAdapter::new_arc() {
-                Ok(lens_adapter) => {
-                    server = server.with_lens_arc(lens_adapter);
-                    info!("Lens HTTP endpoint enabled");
-                }
-                Err(e) => {
-                    warn!("Failed to create lens adapter: {}", e);
-                }
-            }
-
-            // Wire NAC (Node Access Control) to HTTP server only when enabled
-            if config.acp.node_enable {
-                let nac_config = db::NacConfig::new().with_enabled();
-                let nac_manager: std::sync::Arc<dyn db::NacManagerApi> =
-                    std::sync::Arc::new(db::create_memory_nac_manager(nac_config));
-                let nac_adapter = crate::nac_adapter::NacAdapter::new_arc(nac_manager);
-                server = server.with_nac_arc(nac_adapter);
-                info!("NAC HTTP endpoints enabled");
-            } else {
-                info!("NAC disabled (use --acp-node-enable to enable)");
-            }
-
-            // Wire ACP adapters only when document ACP is enabled
-            if config.acp.document_type != AcpDocumentType::None {
-                let acp_adapter = crate::acp_adapter::AcpAdapter::new_arc(zanzibar_store);
-                server = server.with_acp_arc(acp_adapter);
-                info!(
-                    "ACP policy HTTP endpoints enabled (type: {})",
-                    config.acp.document_type
-                );
-
-                let doc_acp_adapter = crate::doc_acp_adapter::DocumentAcpAdapter::new_arc(
-                    database.clone(),
-                    Arc::new(acp::LocalDocumentACP::new(acp_store_for_http)),
-                );
-                server = server.with_doc_acp_arc(doc_acp_adapter);
-                info!("Document ACP HTTP endpoints enabled");
-            } else {
-                info!("Document ACP disabled (use --document-acp-type to enable)");
-            }
-
-            // Wire collection management operations to HTTP server
-            let collection_mgmt_adapter =
-                crate::collection_mgmt_adapter::CollectionManagementAdapter::new_arc(
-                    database.clone(),
-                );
-            server = server.with_collection_mgmt_arc(collection_mgmt_adapter);
-            info!("Collection management HTTP endpoints enabled");
-
-            // Wire index operations to HTTP server
-            let index_adapter = crate::index_adapter::IndexAdapter::new_arc(database.clone());
-            server = server.with_index_arc(index_adapter);
-            info!("Index HTTP endpoints enabled");
-
-            // Wire event bus to HTTP server for GraphQL subscriptions
-            server = server.with_event_bus_arc(event_bus);
-            info!("Subscription event bus enabled");
-
-            info!(
-                "HTTP server configured on {} with REST endpoints enabled",
-                api_address
-            );
-
-            // Build P2PTasks if P2P is enabled with all required task handles
-            let p2p_tasks = match (p2p_host_task, replication_task) {
-                (Some(host_task), Some(replication_task)) => Some(P2PTasks {
-                    host_task,
-                    replication_task,
-                    event_handler_task,
-                }),
-                _ => None,
-            };
-
-            (server, p2p_tasks)
-        };
-
-        Ok((p2p, p2p_tasks, http_server))
-    }
-
-    /// Start P2P networking with the given bitswap store and optional keypair.
-    ///
-    /// If a keypair is provided, it will be used for the P2P identity.
-    /// Otherwise, an ephemeral keypair will be generated.
-    ///
-    /// Returns the handle, events receiver, and host task handle so that the caller
-    /// can connect events to the sync coordinator and track the host task for shutdown.
-    async fn start_p2p<S: p2p::BitswapStore>(
-        config: &Config,
-        bitswap_store: S,
-        keypair: Option<p2p::Keypair>,
-        enable_pubsub: bool,
-    ) -> Result<(
-        p2p::P2PHostHandle,
-        tokio::sync::mpsc::Receiver<p2p::HostEvent>,
-        JoinHandle<()>,
-    )> {
-        let (host, handle, events, _replicators) = match keypair {
-            Some(kp) => p2p::P2PHost::with_keypair_and_config(kp, bitswap_store, enable_pubsub)
-                .await
-                .map_err(Error::P2P)?,
-            None => p2p::P2PHost::with_pubsub(bitswap_store, enable_pubsub)
-                .await
-                .map_err(Error::P2P)?,
-        };
-
-        // Spawn the host event loop FIRST - it must be running to process commands
-        // Track the task handle for graceful shutdown
-        let host_task = tokio::spawn(host.run());
-
-        // Start listening on configured addresses
-        for addr_str in &config.net.p2p_addresses {
-            let addr: p2p::Multiaddr = addr_str
-                .parse()
-                .map_err(|e| Error::InvalidMultiaddr(format!("{}: {}", addr_str, e)))?;
-
-            handle.listen(addr.clone()).await.map_err(Error::P2P)?;
-            info!("P2P listening on {}", addr);
-        }
-
-        // Log bootstrap peers
-        if !config.net.peers.is_empty() {
-            info!("Bootstrap peers configured: {:?}", config.net.peers);
-        }
-
-        if config.net.pubsub_enabled {
-            info!("GossipSub pubsub enabled");
-        } else {
-            info!("GossipSub pubsub disabled");
-        }
-
-        if config.net.relay_enabled {
-            warn!("net.relay_enabled=true is not yet supported; relay is not available");
-        }
-
-        // Get and display peer ID
-        match handle.local_peer_id().await {
-            Ok(peer_id) => info!("Local peer ID: {}", peer_id),
-            Err(e) => error!("Failed to get local peer ID: {}", e),
-        }
-
-        Ok((handle, events, host_task))
-    }
-
-    /// Run the node until shutdown
-    #[doc(hidden)]
-    pub async fn run(mut self) -> Result<()> {
-        info!("DefraDB node started");
-        info!("API endpoint: http://{}", self.config.api.address);
-
-        // Start HTTP server
-        let http_task: Option<JoinHandle<()>> = if let Some(server) = self.http_server.take() {
-            info!("Starting HTTP server on {}", self.config.api.address);
-            Some(tokio::spawn(async move {
-                if let Err(e) = server.run().await {
-                    error!("HTTP server error: {}", e);
-                }
-            }))
-        } else {
-            None
-        };
-
-        // Set up signal handling
-        let shutdown_tx = self.shutdown_tx.clone();
-
-        #[cfg(unix)]
-        {
-            use tokio::signal::unix::{signal, SignalKind};
-
-            let mut sigint =
-                signal(SignalKind::interrupt()).map_err(|e| Error::Signal(e.to_string()))?;
-            let mut sigterm =
-                signal(SignalKind::terminate()).map_err(|e| Error::Signal(e.to_string()))?;
-
-            tokio::spawn(async move {
-                tokio::select! {
-                    _ = sigint.recv() => {
-                        info!("Received SIGINT");
-                    }
-                    _ = sigterm.recv() => {
-                        info!("Received SIGTERM");
-                    }
-                }
-                if let Err(e) = shutdown_tx.send(()).await {
-                    error!("Failed to send shutdown signal: {}", e);
-                }
-            });
-        }
-
-        #[cfg(not(unix))]
-        {
-            tokio::spawn(async move {
-                match tokio::signal::ctrl_c().await {
-                    Ok(()) => {
-                        info!("Received Ctrl+C");
-                        if let Err(e) = shutdown_tx.send(()).await {
-                            error!("Failed to send shutdown signal: {}", e);
-                        }
-                    }
-                    Err(e) => {
-                        error!(
-                            "Failed to listen for Ctrl+C signal: {}. Node may not respond to interrupt signals.",
-                            e
-                        );
-                    }
-                }
-            });
-        }
-
-        // Wait for shutdown signal OR HTTP server crash
-        let mut http_task = http_task;
-        let http_crashed = match &mut http_task {
-            Some(task) => {
-                tokio::select! {
-                    _ = self.shutdown_rx.recv() => false,
-                    result = task => {
-                        match result {
-                            Ok(()) => {
-                                error!("HTTP server exited unexpectedly");
-                            }
-                            Err(e) if e.is_panic() => {
-                                error!("HTTP server panicked: {}", e);
-                            }
-                            Err(e) => {
-                                error!("HTTP server task failed: {}", e);
-                            }
-                        }
-                        true
-                    }
-                }
-            }
-            None => {
-                self.shutdown_rx.recv().await;
-                false
-            }
-        };
-
-        if http_crashed {
-            info!("Initiating shutdown due to HTTP server failure...");
-        } else {
-            info!("Shutting down DefraDB node...");
-            // Only abort if we're shutting down normally (not due to crash)
-            if let Some(task) = http_task {
-                info!("Stopping HTTP server...");
-                task.abort();
-                match tokio::time::timeout(std::time::Duration::from_secs(1), task).await {
-                    Ok(_) => info!("HTTP server stopped"),
-                    Err(_) => warn!(
-                        timeout_secs = 1,
-                        "HTTP server shutdown timed out - server was forcefully terminated. \
-                         This may occur if requests were still in flight."
-                    ),
-                }
-            }
-        }
-
-        // Shutdown P2P tasks first, then the handle
-        // Note: We log but don't return errors here because:
-        // 1. The user's intent (stop the node) is still fulfilled
-        // 2. The node is already stopping - failing would just add noise
-        // 3. P2P cleanup issues don't affect data integrity
-        if let Some(tasks) = self.p2p_tasks.take() {
-            info!("Stopping P2P background tasks...");
-
-            // Abort all tasks - they will stop when the channel closes
-            tasks.replication_task.abort();
-            tasks.host_task.abort();
-            if let Some(event_task) = tasks.event_handler_task {
-                event_task.abort();
-            }
-
-            // Wait briefly for tasks to complete with timeout
-            let timeout = std::time::Duration::from_secs(2);
-            let _ = tokio::time::timeout(timeout, tasks.replication_task).await;
-            let _ = tokio::time::timeout(timeout, tasks.host_task).await;
-
-            info!("P2P background tasks stopped");
-        }
-
-        if let Some(handle) = &self.p2p_handle {
-            if let Err(e) = handle.shutdown().await {
-                warn!("P2P shutdown encountered an issue: {}", e);
-            }
-        }
-
-        info!("DefraDB node shutdown complete");
-        Ok(())
-    }
-}
-
-// Tests extracted to crates/cli/tests/http_integration_tests.rs
