+++ crates/cli/tests/start_tests.rs
@@ -23,6 +23,7 @@ fn default_start_args() -> StartArgs {
         no_searchable_encryption: None,
         identity: None,
         replicator_retry_intervals: None,
+        durability: None,
     }
 }
 
@@ -92,6 +93,7 @@ fn test_apply_to_config_all_flags() {
         no_searchable_encryption: Some(true),
         identity: None, // identity is handled in Node::new, not apply_to_config
         replicator_retry_intervals: Some(vec![10, 20, 30]),
+        durability: None,
     };
 
     let result = args.apply_to_config(&mut config);
